Adding new functions to contract. Adding run.js file for testing
